+++ R/mousecoords.R
@@ -13,6 +13,8 @@
 #' @param epsg the epsg string to be shown.
 #' @param proj4string the proj4string to be shown.
 #' @param native.crs logical. whether to use the native crs in the coordinates box.
+#' @param css list of valid CSS key-value pairs. See e.g.
+#' \url{https://www.w3schools.com/cssref/index.php} for possible values.
 #'
 #' @details
 #' If style is set to "detailed", the following information will be displayed:
@@ -45,6 +47,18 @@
 #'
 #' removeMouseCoordinates(m)
 #'
+#' # adjust css
+#' leaflet() %>%
+#'   addProviderTiles("OpenStreetMap") %>%
+#'   addMouseCoordinates(
+#'     css = list(
+#'      "font-size" = "30px",
+#'      "text-align" = "center",
+#'      "background-color" = "#ff000080",
+#'      "color" = "rgb(255, 255, 255)"
+#'    )
+#'  )
+#'
 #' @export addMouseCoordinates
 #' @name addMouseCoordinates
 #' @rdname addMouseCoordinates
@@ -53,7 +67,8 @@
 addMouseCoordinates <- function(map,
                                 epsg = NULL,
                                 proj4string = NULL,
-                                native.crs = FALSE) {
+                                native.crs = FALSE,
+                                css = list()) {
 
   if (inherits(map, "mapview")) map <- mapview2leaflet(map)
   stopifnot(inherits(map, c("leaflet", "leaflet_proxy", "mapdeck")))
@@ -86,6 +101,24 @@ addMouseCoordinates <- function(map,
     clipboardDependency()
   )
 
+  css_dflt = list(
+    'position' = 'relative'
+    , 'bottomleft' =  '0px'
+    , 'background-color' = 'rgba(255, 255, 255, 0.7)'
+    , 'box-shadow' = '0 0 2px #bbb'
+    , 'background-clip' = 'padding-box'
+    , 'margin' = '0'
+    , 'padding-left' = '5px'
+    , 'padding-right' = '5px'
+    , 'color' = '#333'
+    , 'font-size' = '9px'
+    , 'font-family' = '\"Helvetica Neue\", Arial, Helvetica, sans-serif'
+    , 'text-align' = 'left'
+    , 'z-index' = '700'
+  )
+
+  css = utils::modifyList(css_dflt, css)
+
   map <- htmlwidgets::onRender(
     map,
     paste0(
@@ -103,18 +136,7 @@ addMouseCoordinates <- function(map,
       $(el).append(newDiv);
       //provide ID and style
       newDiv.addClass('lnlt');
-      newDiv.css({
-      'position': 'relative',
-      'bottomleft':  '0px',
-      'background-color': 'rgba(255, 255, 255, 0.7)',
-      'box-shadow': '0 0 2px #bbb',
-      'background-clip': 'padding-box',
-      'margin': '0',
-      'padding-left': '5px',
-      'color': '#333',
-      'font': '9px/1.5 \"Helvetica Neue\", Arial, Helvetica, sans-serif',
-      'z-index': '700',
-      });
+      newDiv.css(", jsonlite::toJSON(css, auto_unbox = TRUE), ");
       return newDiv;
       }
 
@@ -243,102 +265,3 @@ mouseCoordsDependenciesMD = function() {
     )
   )
 }
-
-
-#   if (native.crs) { # | map$x$options$crs$crsClass == "L.CRS.Simple") {
-#     txt_detailed <- paste0("
-#                            ' x: ' + (e.latlng.lng).toFixed(5) +
-#                            ' | y: ' + (e.latlng.lat).toFixed(5) +
-#                            ' | epsg: ", epsg, " ' +
-#                            ' | proj4: ", proj4string, " ' +
-#                            ' | zoom: ' + map.getZoom() + ' '")
-#   } else {
-#     txt_detailed <- paste0("
-#                            ' lon: ' + (e.latLng.lng).toFixed(5) +
-#                            ' | lat: ' + (e.latLng.lat).toFixed(5) +
-#                            ' | zoom: ' + map.getZoom() +
-#                            ' | x: ' + L.CRS.EPSG3857.project(e.latlng).x.toFixed(0) +
-#                            ' | y: ' + L.CRS.EPSG3857.project(e.latlng).y.toFixed(0) +
-#                            ' | epsg: 3857 ' +
-#                            ' | proj4: +proj=merc +a=6378137 +b=6378137 +lat_ts=0.0 +lon_0=0.0 +x_0=0.0 +y_0=0 +k=1.0 +units=m +nadgrids=@null +no_defs '")
-#   }
-#
-#   txt_basic <- paste0("
-#                       ' lon: ' + (e.latLng.lng).toFixed(5) +
-#                       ' | lat: ' + (e.latLng.lat).toFixed(5) +
-#                       ' | zoom: ' + map.getZoom() + ' '")
-#
-#   map$dependencies = c(
-#     map$dependencies,
-#     clipboardDependency()
-#   )
-#
-#   map <- htmlwidgets::onRender(
-#     map,
-#     paste0(
-#       "
-#       function(el, x, data) {
-#       // get the leaflet map
-#       //var map = document.querySelector('.mapdeckmap');
-#       var map = window[el.id+'map']._map.getMap();
-#       console.log(map);
-#       // we need a new div element because we have to handle
-#       // the mouseover output separately
-#       function addElement () {
-#       // generate new div Element
-#       var newDiv = document.createElement('div');
-#       // append at end of leaflet htmlwidget container
-#       el.append(newDiv);
-#       //provide ID and style
-#       newDiv.classList.add('lnlt');
-#       newDiv.style.cssText = 'position: relative; bottomleft: 0px; background-color: rgba(255, 255, 255, 0.7); box-shadow: 0 0 2px #bbb; background-clip: padding-box; margin: 0; padding-left: 5px; color: #333; font: 9px/1.5 \"Helvetica Neue\", Arial, Helvetica, sans-serif; z-index: 700; height: 10px;';
-#       return newDiv;
-#       }
-#
-#
-#       // check for already existing lnlt class to not duplicate
-#       var lnlt = document.querySelector('.lnlt');
-#
-#       if(lnlt === null) {
-#       lnlt = addElement();
-#
-#       // grab the special div we generated in the beginning
-#       // and put the mousmove output there
-#
-#       map.on('mousemove', function (e) {
-#       if (e.originalEvent.ctrlKey) {
-#       if (document.querySelector('.lnlt') === null) lnlt = addElement();
-#       lnlt.innerText(", txt_detailed, ");
-#       } else {
-#       if (document.querySelector('.lnlt') === null) lnlt = addElement();
-#       lnlt.innerText(", txt_basic, ");
-#       }
-#       });
-#
-#       // remove the lnlt div when mouse leaves map
-#       map.on('mouseout', function (e) {
-#       var strip = document.querySelector('.lnlt');
-#       strip.remove();
-#       });
-#
-#       };
-#
-#       //$(el).keypress(67, function(e) {
-#       map.on('preclick', function(e) {
-#       if (e.originalEvent.ctrlKey) {
-#       if (document.querySelector('.lnlt') === null) lnlt = addElement();
-#       lnlt.text(", txt_basic, ");
-#       var txt = document.querySelector('.lnlt').textContent;
-#       console.log(txt);
-#       //txt.innerText.focus();
-#       //txt.select();
-#       setClipboardText('\"' + txt + '\"');
-#       }
-#       });
-#
-#       }
-#       "
-#     )
-#   )
-#   map
-# }