+++ NEWS.md
@@ -1,4 +1,16 @@
+# leafem 0.2.4 (2025-04-30)
+
+#### ✨ features and improvements
+
+  * `addMouseCoordinates()` gained argument `css` to modify the layout of the div. #95
+  * `addHomeButton()` gained arguments `css` and `hover_css` to modify the layout. #92
+  * `addRasterRGB()` and `addStarsImage()` now support more interpolation options. #106
+
+
 # leafem 0.2.3 (2023-09-17)
+
+  * fix CRAN Note (character too long)
+
 
 # leafem 0.2.2 (2023-09-16)
 
@@ -80,4 +92,4 @@ miscellaneous:
 
 ## leafem 0.0.1
 
-initial commit
+  * initial commit